Allow multiple xml files
