+++ src/AzureFunctions.Extensions.Swashbuckle/AzureFunctions.Extensions.Swashbuckle.sln
@@ -7,6 +7,8 @@ Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "AzureFunctions.Extensions.S
 EndProject
 Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "TestFunction", "TestFunction\TestFunction.csproj", "{93B84BC0-2017-4BC2-8191-20EDF594E7AC}"
 EndProject
+Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "TestModels", "TestModels\TestModels.csproj", "{FD596D76-15C4-4522-A2ED-909A7D2337B7}"
+EndProject
 Global
 	GlobalSection(SolutionConfigurationPlatforms) = preSolution
 		Debug|Any CPU = Debug|Any CPU
@@ -21,6 +23,10 @@ Global
 		{93B84BC0-2017-4BC2-8191-20EDF594E7AC}.Debug|Any CPU.Build.0 = Debug|Any CPU
 		{93B84BC0-2017-4BC2-8191-20EDF594E7AC}.Release|Any CPU.ActiveCfg = Release|Any CPU
 		{93B84BC0-2017-4BC2-8191-20EDF594E7AC}.Release|Any CPU.Build.0 = Release|Any CPU
+		{FD596D76-15C4-4522-A2ED-909A7D2337B7}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
+		{FD596D76-15C4-4522-A2ED-909A7D2337B7}.Debug|Any CPU.Build.0 = Debug|Any CPU
+		{FD596D76-15C4-4522-A2ED-909A7D2337B7}.Release|Any CPU.ActiveCfg = Release|Any CPU
+		{FD596D76-15C4-4522-A2ED-909A7D2337B7}.Release|Any CPU.Build.0 = Release|Any CPU
 	EndGlobalSection
 	GlobalSection(SolutionProperties) = preSolution
 		HideSolutionNode = FALSE
