+++ src/AzureFunctions.Extensions.Swashbuckle/TestFunction/TestFunction.csproj
@@ -4,13 +4,14 @@
     <AzureFunctionsVersion>v2</AzureFunctionsVersion>
   </PropertyGroup>
   <PropertyGroup Condition="'$(Configuration)|$(Platform)'=='Debug|AnyCPU'">
-    <DocumentationFile>C:\Users\yukaa\source\repos\Swashbuckle.WebJobs.Extensions\TestFunction\TestFunction.xml</DocumentationFile>
+    <DocumentationFile>TestFunction.xml</DocumentationFile>
   </PropertyGroup>
   <ItemGroup>
     <PackageReference Include="Microsoft.NET.Sdk.Functions" Version="1.0.24" />
   </ItemGroup>
   <ItemGroup>
     <ProjectReference Include="..\AzureFunctions.Extensions.Swashbuckle\AzureFunctions.Extensions.Swashbuckle.csproj" />
+    <ProjectReference Include="..\TestModels\TestModels.csproj" />
   </ItemGroup>
   <ItemGroup>
     <None Update="host.json">
