+++ src/AzureFunctions.Extensions.Swashbuckle/TestFunction/host.json
@@ -2,7 +2,7 @@
   "version": "2.0",
   "extensions": {
     "Swashbuckle": {
-      "XmlPath":  "TestFunction.xml" 
-    } 
+      "XmlPath": "TestFunction.xml;TestModels.xml"
+    }
   }
 }
